Set up left drawer for help related/FAQ questions
Set up left side toggleable menu for FAQ questions. Did not update it
yet as I have not decided how I want to create it.. Goal is to make it
as dynamic as possible so FAQ can be updated without having to rebuild.
Unsure yet. Also set up test button in center of mainwindow. Aim to
replace this with rotating circle of buttons for ecah client type that
we want to generate  asettlement for.